+++ include/gcb/ai/ScriptedTask.h
@@ -61,7 +61,7 @@ public:
 	virtual tcGameStream& operator>>(tcGameStream& stream);
 
     ScriptedTask(std::shared_ptr<tcPlatformObject> platform_, Blackboard* bb, 
-        long id_, double priority_, int attributes_, const std::string& scriptName_);
+        int id_, double priority_, int attributes_, const std::string& scriptName_);
     ~ScriptedTask();
     
 protected:
