+++ include/gcb/ai/SelfPreservation.h
@@ -57,7 +57,7 @@ public:
 	virtual tcGameStream& operator>>(tcGameStream& stream);
 
     SelfPreservation(std::shared_ptr<tcPlatformObject> platform_, Blackboard* bb, 
-        long id_, double priority_, int attributes_, const std::string& taskName_);
+        int id_, double priority_, int attributes_, const std::string& taskName_);
     ~SelfPreservation();
     
 private:
