+++ include/gcb/database/md5.h
@@ -43,7 +43,7 @@ typedef unsigned short int UINT2;
 
 #ifndef UINT4
 /* UINT4 defines a four byte word */
-typedef unsigned long int UINT4;
+typedef unsigned int UINT4;
 #endif
